+++ src/ContaTerminal.java
@@ -1,7 +1,4 @@
-import java.text.DecimalFormat;
-import java.text.DecimalFormatSymbols;
-import java.util.Locale;
-import java.util.Scanner;
+import MenuInicial.MenuInicial;
 
 /**
  * @author Gabriel Santos