+++ RELEASENOTES.md
@@ -5,7 +5,23 @@
 > - [Demonstração da Versão 1.0.0](https://drive.google.com/file/d/1p6kFwsWTDmtFIXkBAmxeerPapEh2VGDR/view?usp=sharing "Demonstração de como o sistema está sendo executado na versão (1.0.0).")
 
 ### **1.1.0**:
-> - Adicionado class MenuInicial
-> - Cadastro de cliente da versão _1.0.0_ enviado para dentro da class MenuInicial
+> - Adicionado class MenuInicial.MenuInicial
+> - Cadastro de cliente da versão _1.0.0_ enviado para dentro da class MenuInicial.MenuInicial
 > - Adicionado Opção de Sair
 > - [Demonstração da Versão 1.1.0](https://drive.google.com/file/d/1TBhGc5zXveLf9mNzZmUOKF1-1kgw6uxC/view?usp=sharing "Demonstração de como o sistema está sendo executado na versão (1.1.0).")
+
+### **1.1.1**:
+> - Adicionado class de Cadastro De Cliente.
+> - Retirado Cadastro antigo.
+> - Ajuste na função de carregamento.
+> - Adicionado chamada para a nova class no MenuInicial.MenuInicial.
+> - Adicionado novos dados de cadastro:
+> 
+> | Nome Variavel |  Tipo  | Status |
+> |:-------------:|:------:|:------:|
+> | primeiroNome  | String |   ✅    |
+> |   sobrenome   | String |   ✅    |
+> |     email     | String |   ✅    |
+> |     senha     | String |   ✅    |
+> |      cpf      | String |   ✅    | 
+